🚚 Move: routes/news_slider.svelte -> lib/components/news_slider.svelte
