+++ src/lib/components/news_slider.svelte
